Move loading to different component
